telebot/requirements.txt
python-telegram-bot
pyshorteners
dotenv
requests